Decompile and break stuff
The build is OK now, but I know I've broken it somehow, because code2.s is
getting aligned forward to an 8- or 16-byte boundary for some reason.

To get things working on Linux and ease debugging, I updated a few of the tools
to match pret/pokeemerald. I also got agbcc and libc/libgcc from there, but
they're not used in the build yet afaict so it doesn't matter.